+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master
 
+* Adds support for react-navigation v5
 * Fixes an issue with `onReportSubmitHandler` on iOS
 
 ## v9.1.7 (2020-08-10)